requirements.txt
discord.py==2.2.2
python-dotenv==1.0.0
spotipy==2.22.1
yt-dlp==2023.3.4
PyNaCl==1.5.0